style(cli): ruff format + extract TUI components (AES406)
- ruff format on doctor/controller/tui_app (CI Format gate)
- E501: split long U8 warning line
- Extract FilePickerModal/HelpScreen/QwenTuiLogHandler into
  surface_cli_tui_components.py so the app surface stays under the
  AES406 control-flow budget (Self-Lint gate)
- Remove now-unused imports (escape, ModalScreen, DirectoryTree) from
  the app surface after extraction
