+++ modules/cli/src/surface_cli_interactive_controller.py
@@ -55,7 +55,7 @@ class InteractiveController:
                     "Interactive mode requires a TTY, but stdin is not interactive (pipe/cron detected).\n"
                     "Why: the Obsidian Nebula TUI needs a real terminal to render.\n"
                     "How to fix: use a non-interactive subcommand instead, e.g.:\n"
-                    "  qwen-web-cli prompt-direct -t \"Summarize this\" [--json]\n"
+                    '  qwen-web-cli prompt-direct -t "Summarize this" [--json]\n'
                     "  qwen-web-cli prompt-only -i prompt.md [--json]\n"
                     "  qwen-web-cli prompt-with-attachment -i prompt.md -a report.pdf [--json]\n"
                     "Then verify your environment with: qwen-web-cli doctor"
